Clean up the client context a bit We're not ready to make use of most of these options
